pyproject.toml
[build-system] # Tells python how to build the package
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project] # metadata
name = "pdfharvest"
version = "0.1.0"
description = "Automatic publication downloader and analyzer"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Laura", email = "lsanz@student.agh.edu.pl" }]
license = { text = "MIT" }
dependencies = [
  "httpx>=0.27",
  "PyYAML>=6.0",
  "typer>=0.12",
]

[tool.setuptools.packages.find]  #Tells set up tools to look for code in src/
where = ["src"]

[tool.pytest.ini_options] # Test configuration for pytest
addopts = "-q"
pythonpath = ["src"]

[project.scripts]
pdfharvest = "pdfharvest.cli:cli"